+++ src/App.tsx
@@ -19,9 +19,9 @@ import MyPage from './pages/MyPage';
 
 import OnboardingProfilePage from "./pages/OnboardingProfilePage";
 import OnboardingInterestsPage from "./pages/OnboardingInterestsPage";
+import ProtectedLayer from './components/ProtectedLayer';
 
 function App() {
-  const isLoggedIn = !!localStorage.getItem("token");
 
   return (
     <BrowserRouter>
@@ -31,23 +31,23 @@ function App() {
       <Route path="/login" element={<LoginPage />} />
       <Route path="/signup" element={<SignupPage />} />
       <Route path="/api/auth/kakao/callback" element={<OAuthKakaoCallback />} />
-        <Route path="/calendar" element={<CalendarPage />} />
-        <Route path="/recommend" element={<RecommendPage/>} />
-        <Route path="/recommend/result/:category" element={<RecommandResultTemplate />} />
-        
+      <Route path="/calendar" element={<CalendarPage />} />
+      <Route path="/recommend" element={<RecommendPage/>} />
+      
+      <Route element={<ProtectedLayer />}>
+      <Route path="/recommend/result/:category" element={<RecommandResultTemplate />} />
+      <Route path="/friends" element={<FriendsPage />} />
+      <Route path="/friends/requests" element={<FriendRequestPage />} />
+      <Route path="/friend-profile/:friendId" element={<FriendProfilePage />} />
+      <Route path="/friend/:friendId" element={<FriendProfilePage />} />
+      <Route path="/friends/chat/*" element={<ChatPage />} />
+      <Route path="/friends/search" element={<FriendSearchPage />} />
+      <Route path="/mypage/*" element= {<MyPage/>}/>
+      </Route>  
 
-        <Route path="/friends" element={<FriendsPage />} />
-        <Route path="/friends/requests" element={<FriendRequestPage />} />
-        <Route path="/friend-profile/:friendId" element={<FriendProfilePage />} />
-        <Route path="/friend/:friendId" element={<FriendProfilePage />} />
-        <Route path="/friends/chat/*" element={<ChatPage />} />
-        <Route path="/friends/search" element={<FriendSearchPage />} />
-        <Route path="/mypage/*"
-          element={isLoggedIn ? <MyPage /> : <MyPage/>}/>
-          {/* <Navigate to="/login" /> */}
-        <Route path="/onboarding/profile" element={<OnboardingProfilePage />} />
-        <Route path="/onboarding/interests" element={<OnboardingInterestsPage />} />
-        <Route path="/onboarding" element={<Navigate to="/onboarding/profile" replace />} />
+      <Route path="/onboarding/profile" element={<OnboardingProfilePage />} />
+      <Route path="/onboarding/interests" element={<OnboardingInterestsPage />} />
+      <Route path="/onboarding" element={<Navigate to="/onboarding/profile" replace />} />
       </Routes>
       <ToastContainer position="bottom-right" autoClose={1500} hideProgressBar newestOnTop closeOnClick pauseOnFocusLoss={false} draggable pauseOnHover theme="light"/>
     </BrowserRouter>
